deps/v8/test/webkit/dfg-put-by-id-prototype-check-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Tests that the DFG CFA is not overzealous in removing prototype structure checks for put_by_id transitions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS o.f is 0
PASS o.f is 1
PASS o.f is 2
PASS o.f is 3
PASS o.f is 4
PASS o.f is 5
PASS o.f is 6
PASS o.f is 7
PASS o.f is 8
PASS o.f is 9
PASS o.f is 10
PASS o.f is 11
PASS o.f is 12
PASS o.f is 13
PASS o.f is 14
PASS o.f is 15
PASS o.f is 16
PASS o.f is 17
PASS o.f is 18
PASS o.f is 19
PASS o.f is 20
PASS o.f is 21
PASS o.f is 22
PASS o.f is 23
PASS o.f is 24
PASS o.f is 25
PASS o.f is 26
PASS o.f is 27
PASS o.f is 28
PASS o.f is 29
PASS o.f is 30
PASS o.f is 31
PASS o.f is 32
PASS o.f is 33
PASS o.f is 34
PASS o.f is 35
PASS o.f is 36
PASS o.f is 37
PASS o.f is 38
PASS o.f is 39
PASS o.f is 40
PASS o.f is 41
PASS o.f is 42
PASS o.f is 43
PASS o.f is 44
PASS o.f is 45
PASS o.f is 46
PASS o.f is 47
PASS o.f is 48
PASS o.f is 49
PASS o.f is 50
PASS o.f is 51
PASS o.f is 52
PASS o.f is 53
PASS o.f is 54
PASS o.f is 55
PASS o.f is 56
PASS o.f is 57
PASS o.f is 58
PASS o.f is 59
PASS o.f is 60
PASS o.f is 61
PASS o.f is 62
PASS o.f is 63
PASS o.f is 64
PASS o.f is 65
PASS o.f is 66
PASS o.f is 67
PASS o.f is 68
PASS o.f is 69
PASS o.f is 70
PASS o.f is 71
PASS o.f is 72
PASS o.f is 73
PASS o.f is 74
PASS o.f is 75
PASS o.f is 76
PASS o.f is 77
PASS o.f is 78
PASS o.f is 79
PASS o.f is 80
PASS o.f is 81
PASS o.f is 82
PASS o.f is 83
PASS o.f is 84
PASS o.f is 85
PASS o.f is 86
PASS o.f is 87
PASS o.f is 88
PASS o.f is 89
PASS o.f is 90
PASS o.f is 91
PASS o.f is 92
PASS o.f is 93
PASS o.f is 94
PASS o.f is 95
PASS o.f is 96
PASS o.f is 97
PASS o.f is 98
PASS o.f is 99
PASS o.f is 100
PASS o.f is 101
PASS o.f is 102
PASS o.f is 103
PASS o.f is 104
PASS o.f is 105
PASS o.f is 106
PASS o.f is 107
PASS o.f is 108
PASS o.f is 109
PASS o.f is 110
PASS o.f is 111
PASS o.f is 112
PASS o.f is 113
PASS o.f is 114
PASS o.f is 115
PASS o.f is 116
PASS o.f is 117
PASS o.f is 118
PASS o.f is 119
PASS o.f is 120
PASS o.f is 121
PASS o.f is 122
PASS o.f is 123
PASS o.f is 124
PASS o.f is 125
PASS o.f is 126
PASS o.f is 127
PASS o.f is 128
PASS o.f is 129
PASS o.f is 130
PASS o.f is 131
PASS o.f is 132
PASS o.f is 133
PASS o.f is 134
PASS o.f is 135
PASS o.f is 136
PASS o.f is 137
PASS o.f is 138
PASS o.f is 139
PASS o.f is 140
PASS o.f is 141
PASS o.f is 142
PASS o.f is 143
PASS o.f is 144
PASS o.f is 145
PASS o.f is 146
PASS o.f is 147
PASS o.f is 148
PASS o.f is 149
PASS o.f is 150
PASS o.f is 151
PASS o.f is 152
PASS o.f is 153
PASS o.f is 154
PASS o.f is 155
PASS o.f is 156
PASS o.f is 157
PASS o.f is 158
PASS o.f is 159
PASS o.f is 160
PASS o.f is 161
PASS o.f is 162
PASS o.f is 163
PASS o.f is 164
PASS o.f is 165
PASS o.f is 166
PASS o.f is 167
PASS o.f is 168
PASS o.f is 169
PASS o.f is 170
PASS o.f is 171
PASS o.f is 172
PASS o.f is 173
PASS o.f is 174
PASS o.f is 175
PASS o.f is 176
PASS o.f is 177
PASS o.f is 178
PASS o.f is 179
PASS o.f is 180
PASS o.f is 181
PASS o.f is 182
PASS o.f is 183
PASS o.f is 184
PASS o.f is 185
PASS o.f is 186
PASS o.f is 187
PASS o.f is 188
PASS o.f is 189
PASS o.f is 190
PASS o.f is 191
PASS o.f is 192
PASS o.f is 193
PASS o.f is 194
PASS o.f is 195
PASS o.f is 196
PASS o.f is 197
PASS o.f is 198
PASS o.f is 199
PASS o.f is 200
PASS o.f is 201
PASS o.f is 202
PASS o.f is 203
PASS o.f is 204
PASS o.f is 205
PASS o.f is 206
PASS o.f is 207
PASS o.f is 208
PASS o.f is 209
PASS o.f is 210
PASS o.f is 211
PASS o.f is 212
PASS o.f is 213
PASS o.f is 214
PASS o.f is 215
PASS o.f is 216
PASS o.f is 217
PASS o.f is 218
PASS o.f is 219
PASS o.f is 220
PASS o.f is 221
PASS o.f is 222
PASS o.f is 223
PASS o.f is 224
PASS o.f is 225
PASS o.f is 226
PASS o.f is 227
PASS o.f is 228
PASS o.f is 229
PASS o.f is 230
PASS o.f is 231
PASS o.f is 232
PASS o.f is 233
PASS o.f is 234
PASS o.f is 235
PASS o.f is 236
PASS o.f is 237
PASS o.f is 238
PASS o.f is 239
PASS o.f is 240
PASS o.f is 241
PASS o.f is 242
PASS o.f is 243
PASS o.f is 244
PASS o.f is 245
PASS o.f is 246
PASS o.f is 247
PASS o.f is 248
PASS o.f is 249
PASS o.f is 250
PASS o.f is 251
PASS o.f is 252
PASS o.f is 253
PASS o.f is 254
PASS o.f is 255
PASS o.f is 256
PASS o.f is 257
PASS o.f is 258
PASS o.f is 259
PASS o.f is 260
PASS o.f is 261
PASS o.f is 262
PASS o.f is 263
PASS o.f is 264
PASS o.f is 265
PASS o.f is 266
PASS o.f is 267
PASS o.f is 268
PASS o.f is 269
PASS o.f is 270
PASS o.f is 271
PASS o.f is 272
PASS o.f is 273
PASS o.f is 274
PASS o.f is 275
PASS o.f is 276
PASS o.f is 277
PASS o.f is 278
PASS o.f is 279
PASS o.f is 280
PASS o.f is 281
PASS o.f is 282
PASS o.f is 283
PASS o.f is 284
PASS o.f is 285
PASS o.f is 286
PASS o.f is 287
PASS o.f is 288
PASS o.f is 289
PASS o.f is 290
PASS o.f is 291
PASS o.f is 292
PASS o.f is 293
PASS o.f is 294
PASS o.f is 295
PASS o.f is 296
PASS o.f is 297
PASS o.f is 298
PASS o.f is 299
PASS o.f is 300
PASS o.f is 301
PASS o.f is 302
PASS o.f is 303
PASS o.f is 304
PASS o.f is 305
PASS o.f is 306
PASS o.f is 307
PASS o.f is 308
PASS o.f is 309
PASS o.f is 310
PASS o.f is 311
PASS o.f is 312
PASS o.f is 313
PASS o.f is 314
PASS o.f is 315
PASS o.f is 316
PASS o.f is 317
PASS o.f is 318
PASS o.f is 319
PASS o.f is 320
PASS o.f is 321
PASS o.f is 322
PASS o.f is 323
PASS o.f is 324
PASS o.f is 325
PASS o.f is 326
PASS o.f is 327
PASS o.f is 328
PASS o.f is 329
PASS o.f is 330
PASS o.f is 331
PASS o.f is 332
PASS o.f is 333
PASS o.f is 334
PASS o.f is 335
PASS o.f is 336
PASS o.f is 337
PASS o.f is 338
PASS o.f is 339
PASS o.f is 340
PASS o.f is 341
PASS o.f is 342
PASS o.f is 343
PASS o.f is 344
PASS o.f is 345
PASS o.f is 346
PASS o.f is 347
PASS o.f is 348
PASS o.f is 349
PASS o.f is 350
PASS o.f is 351
PASS o.f is 352
PASS o.f is 353
PASS o.f is 354
PASS o.f is 355
PASS o.f is 356
PASS o.f is 357
PASS o.f is 358
PASS o.f is 359
PASS o.f is 360
PASS o.f is 361
PASS o.f is 362
PASS o.f is 363
PASS o.f is 364
PASS o.f is 365
PASS o.f is 366
PASS o.f is 367
PASS o.f is 368
PASS o.f is 369
PASS o.f is 370
PASS o.f is 371
PASS o.f is 372
PASS o.f is 373
PASS o.f is 374
PASS o.f is 375
PASS o.f is 376
PASS o.f is 377
PASS o.f is 378
PASS o.f is 379
PASS o.f is 380
PASS o.f is 381
PASS o.f is 382
PASS o.f is 383
PASS o.f is 384
PASS o.f is 385
PASS o.f is 386
PASS o.f is 387
PASS o.f is 388
PASS o.f is 389
PASS o.f is 390
PASS o.f is 391
PASS o.f is 392
PASS o.f is 393
PASS o.f is 394
PASS o.f is 395
PASS o.f is 396
PASS o.f is 397
PASS o.f is 398
PASS o.f is 399
PASS o.f is 400
PASS o.f is 401
PASS o.f is 402
PASS o.f is 403
PASS o.f is 404
PASS o.f is 405
PASS o.f is 406
PASS o.f is 407
PASS o.f is 408
PASS o.f is 409
PASS o.f is 410
PASS o.f is 411
PASS o.f is 412
PASS o.f is 413
PASS o.f is 414
PASS o.f is 415
PASS o.f is 416
PASS o.f is 417
PASS o.f is 418
PASS o.f is 419
PASS o.f is 420
PASS o.f is 421
PASS o.f is 422
PASS o.f is 423
PASS o.f is 424
PASS o.f is 425
PASS o.f is 426
PASS o.f is 427
PASS o.f is 428
PASS o.f is 429
PASS o.f is 430
PASS o.f is 431
PASS o.f is 432
PASS o.f is 433
PASS o.f is 434
PASS o.f is 435
PASS o.f is 436
PASS o.f is 437
PASS o.f is 438
PASS o.f is 439
PASS o.f is 440
PASS o.f is 441
PASS o.f is 442
PASS o.f is 443
PASS o.f is 444
PASS o.f is 445
PASS o.f is 446
PASS o.f is 447
PASS o.f is 448
PASS o.f is 449
PASS o.f is 450
PASS o.f is 451
PASS o.f is 452
PASS o.f is 453
PASS o.f is 454
PASS o.f is 455
PASS o.f is 456
PASS o.f is 457
PASS o.f is 458
PASS o.f is 459
PASS o.f is 460
PASS o.f is 461
PASS o.f is 462
PASS o.f is 463
PASS o.f is 464
PASS o.f is 465
PASS o.f is 466
PASS o.f is 467
PASS o.f is 468
PASS o.f is 469
PASS o.f is 470
PASS o.f is 471
PASS o.f is 472
PASS o.f is 473
PASS o.f is 474
PASS o.f is 475
PASS o.f is 476
PASS o.f is 477
PASS o.f is 478
PASS o.f is 479
PASS o.f is 480
PASS o.f is 481
PASS o.f is 482
PASS o.f is 483
PASS o.f is 484
PASS o.f is 485
PASS o.f is 486
PASS o.f is 487
PASS o.f is 488
PASS o.f is 489
PASS o.f is 490
PASS o.f is 491
PASS o.f is 492
PASS o.f is 493
PASS o.f is 494
PASS o.f is 495
PASS o.f is 496
PASS o.f is 497
PASS o.f is 498
PASS o.f is 499
PASS o.f is 500
PASS o.f is 501
PASS o.f is 502
PASS o.f is 503
PASS o.f is 504
PASS o.f is 505
PASS o.f is 506
PASS o.f is 507
PASS o.f is 508
PASS o.f is 509
PASS o.f is 510
PASS o.f is 511
PASS o.f is 512
PASS o.f is 513
PASS o.f is 514
PASS o.f is 515
PASS o.f is 516
PASS o.f is 517
PASS o.f is 518
PASS o.f is 519
PASS o.f is 520
PASS o.f is 521
PASS o.f is 522
PASS o.f is 523
PASS o.f is 524
PASS o.f is 525
PASS o.f is 526
PASS o.f is 527
PASS o.f is 528
PASS o.f is 529
PASS o.f is 530
PASS o.f is 531
PASS o.f is 532
PASS o.f is 533
PASS o.f is 534
PASS o.f is 535
PASS o.f is 536
PASS o.f is 537
PASS o.f is 538
PASS o.f is 539
PASS o.f is 540
PASS o.f is 541
PASS o.f is 542
PASS o.f is 543
PASS o.f is 544
PASS o.f is 545
PASS o.f is 546
PASS o.f is 547
PASS o.f is 548
PASS o.f is 549
PASS o.f is 550
PASS o.f is 551
PASS o.f is 552
PASS o.f is 553
PASS o.f is 554
PASS o.f is 555
PASS o.f is 556
PASS o.f is 557
PASS o.f is 558
PASS o.f is 559
PASS o.f is 560
PASS o.f is 561
PASS o.f is 562
PASS o.f is 563
PASS o.f is 564
PASS o.f is 565
PASS o.f is 566
PASS o.f is 567
PASS o.f is 568
PASS o.f is 569
PASS o.f is 570
PASS o.f is 571
PASS o.f is 572
PASS o.f is 573
PASS o.f is 574
PASS o.f is 575
PASS o.f is 576
PASS o.f is 577
PASS o.f is 578
PASS o.f is 579
PASS o.f is 580
PASS o.f is 581
PASS o.f is 582
PASS o.f is 583
PASS o.f is 584
PASS o.f is 585
PASS o.f is 586
PASS o.f is 587
PASS o.f is 588
PASS o.f is 589
PASS o.f is 590
PASS o.f is 591
PASS o.f is 592
PASS o.f is 593
PASS o.f is 594
PASS o.f is 595
PASS o.f is 596
PASS o.f is 597
PASS o.f is 598
PASS o.f is 599
PASS o.f is 600
PASS o.f is 601
PASS o.f is 602
PASS o.f is 603
PASS o.f is 604
PASS o.f is 605
PASS o.f is 606
PASS o.f is 607
PASS o.f is 608
PASS o.f is 609
PASS o.f is 610
PASS o.f is 611
PASS o.f is 612
PASS o.f is 613
PASS o.f is 614
PASS o.f is 615
PASS o.f is 616
PASS o.f is 617
PASS o.f is 618
PASS o.f is 619
PASS o.f is 620
PASS o.f is 621
PASS o.f is 622
PASS o.f is 623
PASS o.f is 624
PASS o.f is 625
PASS o.f is 626
PASS o.f is 627
PASS o.f is 628
PASS o.f is 629
PASS o.f is 630
PASS o.f is 631
PASS o.f is 632
PASS o.f is 633
PASS o.f is 634
PASS o.f is 635
PASS o.f is 636
PASS o.f is 637
PASS o.f is 638
PASS o.f is 639
PASS o.f is 640
PASS o.f is 641
PASS o.f is 642
PASS o.f is 643
PASS o.f is 644
PASS o.f is 645
PASS o.f is 646
PASS o.f is 647
PASS o.f is 648
PASS o.f is 649
PASS o.f is 650
PASS o.f is 651
PASS o.f is 652
PASS o.f is 653
PASS o.f is 654
PASS o.f is 655
PASS o.f is 656
PASS o.f is 657
PASS o.f is 658
PASS o.f is 659
PASS o.f is 660
PASS o.f is 661
PASS o.f is 662
PASS o.f is 663
PASS o.f is 664
PASS o.f is 665
PASS o.f is 666
PASS o.f is 667
PASS o.f is 668
PASS o.f is 669
PASS o.f is 670
PASS o.f is 671
PASS o.f is 672
PASS o.f is 673
PASS o.f is 674
PASS o.f is 675
PASS o.f is 676
PASS o.f is 677
PASS o.f is 678
PASS o.f is 679
PASS o.f is 680
PASS o.f is 681
PASS o.f is 682
PASS o.f is 683
PASS o.f is 684
PASS o.f is 685
PASS o.f is 686
PASS o.f is 687
PASS o.f is 688
PASS o.f is 689
PASS o.f is 690
PASS o.f is 691
PASS o.f is 692
PASS o.f is 693
PASS o.f is 694
PASS o.f is 695
PASS o.f is 696
PASS o.f is 697
PASS o.f is 698
PASS o.f is 699
PASS o.f is 700
PASS o.f is 701
PASS o.f is 702
PASS o.f is 703
PASS o.f is 704
PASS o.f is 705
PASS o.f is 706
PASS o.f is 707
PASS o.f is 708
PASS o.f is 709
PASS o.f is 710
PASS o.f is 711
PASS o.f is 712
PASS o.f is 713
PASS o.f is 714
PASS o.f is 715
PASS o.f is 716
PASS o.f is 717
PASS o.f is 718
PASS o.f is 719
PASS o.f is 720
PASS o.f is 721
PASS o.f is 722
PASS o.f is 723
PASS o.f is 724
PASS o.f is 725
PASS o.f is 726
PASS o.f is 727
PASS o.f is 728
PASS o.f is 729
PASS o.f is 730
PASS o.f is 731
PASS o.f is 732
PASS o.f is 733
PASS o.f is 734
PASS o.f is 735
PASS o.f is 736
PASS o.f is 737
PASS o.f is 738
PASS o.f is 739
PASS o.f is 740
PASS o.f is 741
PASS o.f is 742
PASS o.f is 743
PASS o.f is 744
PASS o.f is 745
PASS o.f is 746
PASS o.f is 747
PASS o.f is 748
PASS o.f is 749
PASS o.f is 750
PASS o.f is 751
PASS o.f is 752
PASS o.f is 753
PASS o.f is 754
PASS o.f is 755
PASS o.f is 756
PASS o.f is 757
PASS o.f is 758
PASS o.f is 759
PASS o.f is 760
PASS o.f is 761
PASS o.f is 762
PASS o.f is 763
PASS o.f is 764
PASS o.f is 765
PASS o.f is 766
PASS o.f is 767
PASS o.f is 768
PASS o.f is 769
PASS o.f is 770
PASS o.f is 771
PASS o.f is 772
PASS o.f is 773
PASS o.f is 774
PASS o.f is 775
PASS o.f is 776
PASS o.f is 777
PASS o.f is 778
PASS o.f is 779
PASS o.f is 780
PASS o.f is 781
PASS o.f is 782
PASS o.f is 783
PASS o.f is 784
PASS o.f is 785
PASS o.f is 786
PASS o.f is 787
PASS o.f is 788
PASS o.f is 789
PASS o.f is 790
PASS o.f is 791
PASS o.f is 792
PASS o.f is 793
PASS o.f is 794
PASS o.f is 795
PASS o.f is 796
PASS o.f is 797
PASS o.f is 798
PASS o.f is 799
PASS o.f is 800
PASS o.f is 801
PASS o.f is 802
PASS o.f is 803
PASS o.f is 804
PASS o.f is 805
PASS o.f is 806
PASS o.f is 807
PASS o.f is 808
PASS o.f is 809
PASS o.f is 810
PASS o.f is 811
PASS o.f is 812
PASS o.f is 813
PASS o.f is 814
PASS o.f is 815
PASS o.f is 816
PASS o.f is 817
PASS o.f is 818
PASS o.f is 819
PASS o.f is 820
PASS o.f is 821
PASS o.f is 822
PASS o.f is 823
PASS o.f is 824
PASS o.f is 825
PASS o.f is 826
PASS o.f is 827
PASS o.f is 828
PASS o.f is 829
PASS o.f is 830
PASS o.f is 831
PASS o.f is 832
PASS o.f is 833
PASS o.f is 834
PASS o.f is 835
PASS o.f is 836
PASS o.f is 837
PASS o.f is 838
PASS o.f is 839
PASS o.f is 840
PASS o.f is 841
PASS o.f is 842
PASS o.f is 843
PASS o.f is 844
PASS o.f is 845
PASS o.f is 846
PASS o.f is 847
PASS o.f is 848
PASS o.f is 849
PASS o.f is 850
PASS o.f is 851
PASS o.f is 852
PASS o.f is 853
PASS o.f is 854
PASS o.f is 855
PASS o.f is 856
PASS o.f is 857
PASS o.f is 858
PASS o.f is 859
PASS o.f is 860
PASS o.f is 861
PASS o.f is 862
PASS o.f is 863
PASS o.f is 864
PASS o.f is 865
PASS o.f is 866
PASS o.f is 867
PASS o.f is 868
PASS o.f is 869
PASS o.f is 870
PASS o.f is 871
PASS o.f is 872
PASS o.f is 873
PASS o.f is 874
PASS o.f is 875
PASS o.f is 876
PASS o.f is 877
PASS o.f is 878
PASS o.f is 879
PASS o.f is 880
PASS o.f is 881
PASS o.f is 882
PASS o.f is 883
PASS o.f is 884
PASS o.f is 885
PASS o.f is 886
PASS o.f is 887
PASS o.f is 888
PASS o.f is 889
PASS o.f is 890
PASS o.f is 891
PASS o.f is 892
PASS o.f is 893
PASS o.f is 894
PASS o.f is 895
PASS o.f is 896
PASS o.f is 897
PASS o.f is 898
PASS o.f is 899
PASS stuff is 900
PASS o.f is undefined
PASS stuff is 901
PASS o.f is undefined
PASS stuff is 902
PASS o.f is undefined
PASS stuff is 903
PASS o.f is undefined
PASS stuff is 904
PASS o.f is undefined
PASS stuff is 905
PASS o.f is undefined
PASS stuff is 906
PASS o.f is undefined
PASS stuff is 907
PASS o.f is undefined
PASS stuff is 908
PASS o.f is undefined
PASS stuff is 909
PASS o.f is undefined
PASS stuff is 910
PASS o.f is undefined
PASS stuff is 911
PASS o.f is undefined
PASS stuff is 912
PASS o.f is undefined
PASS stuff is 913
PASS o.f is undefined
PASS stuff is 914
PASS o.f is undefined
PASS stuff is 915
PASS o.f is undefined
PASS stuff is 916
PASS o.f is undefined
PASS stuff is 917
PASS o.f is undefined
PASS stuff is 918
PASS o.f is undefined
PASS stuff is 919
PASS o.f is undefined
PASS stuff is 920
PASS o.f is undefined
PASS stuff is 921
PASS o.f is undefined
PASS stuff is 922
PASS o.f is undefined
PASS stuff is 923
PASS o.f is undefined
PASS stuff is 924
PASS o.f is undefined
PASS stuff is 925
PASS o.f is undefined
PASS stuff is 926
PASS o.f is undefined
PASS stuff is 927
PASS o.f is undefined
PASS stuff is 928
PASS o.f is undefined
PASS stuff is 929
PASS o.f is undefined
PASS stuff is 930
PASS o.f is undefined
PASS stuff is 931
PASS o.f is undefined
PASS stuff is 932
PASS o.f is undefined
PASS stuff is 933
PASS o.f is undefined
PASS stuff is 934
PASS o.f is undefined
PASS stuff is 935
PASS o.f is undefined
PASS stuff is 936
PASS o.f is undefined
PASS stuff is 937
PASS o.f is undefined
PASS stuff is 938
PASS o.f is undefined
PASS stuff is 939
PASS o.f is undefined
PASS stuff is 940
PASS o.f is undefined
PASS stuff is 941
PASS o.f is undefined
PASS stuff is 942
PASS o.f is undefined
PASS stuff is 943
PASS o.f is undefined
PASS stuff is 944
PASS o.f is undefined
PASS stuff is 945
PASS o.f is undefined
PASS stuff is 946
PASS o.f is undefined
PASS stuff is 947
PASS o.f is undefined
PASS stuff is 948
PASS o.f is undefined
PASS stuff is 949
PASS o.f is undefined
PASS stuff is 950
PASS o.f is undefined
PASS stuff is 951
PASS o.f is undefined
PASS stuff is 952
PASS o.f is undefined
PASS stuff is 953
PASS o.f is undefined
PASS stuff is 954
PASS o.f is undefined
PASS stuff is 955
PASS o.f is undefined
PASS stuff is 956
PASS o.f is undefined
PASS stuff is 957
PASS o.f is undefined
PASS stuff is 958
PASS o.f is undefined
PASS stuff is 959
PASS o.f is undefined
PASS stuff is 960
PASS o.f is undefined
PASS stuff is 961
PASS o.f is undefined
PASS stuff is 962
PASS o.f is undefined
PASS stuff is 963
PASS o.f is undefined
PASS stuff is 964
PASS o.f is undefined
PASS stuff is 965
PASS o.f is undefined
PASS stuff is 966
PASS o.f is undefined
PASS stuff is 967
PASS o.f is undefined
PASS stuff is 968
PASS o.f is undefined
PASS stuff is 969
PASS o.f is undefined
PASS stuff is 970
PASS o.f is undefined
PASS stuff is 971
PASS o.f is undefined
PASS stuff is 972
PASS o.f is undefined
PASS stuff is 973
PASS o.f is undefined
PASS stuff is 974
PASS o.f is undefined
PASS stuff is 975
PASS o.f is undefined
PASS stuff is 976
PASS o.f is undefined
PASS stuff is 977
PASS o.f is undefined
PASS stuff is 978
PASS o.f is undefined
PASS stuff is 979
PASS o.f is undefined
PASS stuff is 980
PASS o.f is undefined
PASS stuff is 981
PASS o.f is undefined
PASS stuff is 982
PASS o.f is undefined
PASS stuff is 983
PASS o.f is undefined
PASS stuff is 984
PASS o.f is undefined
PASS stuff is 985
PASS o.f is undefined
PASS stuff is 986
PASS o.f is undefined
PASS stuff is 987
PASS o.f is undefined
PASS stuff is 988
PASS o.f is undefined
PASS stuff is 989
PASS o.f is undefined
PASS stuff is 990
PASS o.f is undefined
PASS stuff is 991
PASS o.f is undefined
PASS stuff is 992
PASS o.f is undefined
PASS stuff is 993
PASS o.f is undefined
PASS stuff is 994
PASS o.f is undefined
PASS stuff is 995
PASS o.f is undefined
PASS stuff is 996
PASS o.f is undefined
PASS stuff is 997
PASS o.f is undefined
PASS stuff is 998
PASS o.f is undefined
PASS stuff is 999
PASS o.f is undefined
PASS successfullyParsed is true

TEST COMPLETE
